+++ OJGIR/Cuboid.h
@@ -28,7 +28,7 @@ public:
 	};
 
 	//Cuboid(float x, float y, float z, float dX, float dY, float dZ, float light, float tBRDF);
-	Cuboid(glm::vec3 _pos, glm::vec3 _dim, glm::vec3 _emission, glm::vec3 _brdf, float _P);
+	Cuboid(glm::dvec3 _pos, glm::dvec3 _dim, glm::dvec3 _emission, glm::dvec3 _brdf, double _P);
 
 	void createBuffers();
 	~Cuboid();
